refactor(agw): refactor MConfigLoader to enable test, add tests
This refactoring PR was necessary to enable testing of e.g. #9047.

The general points of refactoring were as follows.

- Removed the Class definition
  - As it was unused and extraneous
- Moved the bulk of logic to a public `load_service_mconfig` function that takes in an istream
  - To allow tests that do not require filesystem read/write and associated complexities
- Moved non-public functions to anonymous namespace
  - Preferred over static declaration that (sometimes) existed

When we have access to Google's abseil library, we should make use of a couple helpers there.

- Consider use of the absl::Status return type instead of Boolean
  - Enables more fidelity in error reporting, error joining, etc
- Shall use absl::Cleanup to ensure closure of opened resources (e.g. the file.close())

In authoring tests, it was discovered that many mal-formed json mconfigs would actually result in an uncaught exception being thrown by our third party json library.  I now catch this exception in the MConfigLoader library and pass back a status instead (failure notice via the boolean return value).

Our CMake project architecture seems to make it ~impossible to access our protobuf definitions for mconfig (in /lte/gateway/..) from the MConfigLoader library that uses them (in /orc8r/gateway/c/..).

In an associated PR, I have a Bazel BUILD file which has no issue with this. Hopefully we'll go the way of Bazel - but until then I've compensated for CMake shortcomings (and directory structure, but ultimately I blame CMake here) - by placing a copy of the mconfig proto compiled outputs (and its dependency) within the MConfigLoader test directory.

This should be blown away as soon as we get off CMake for this library. Or if somebody goes through the large effort to make our CMake more whole.

Signed-off-by: Scott Harrison Moeller <smoeller@fb.com>